[YUNIKORN-3062] Refactor restart_changed_config E2E test suite structure
